refactor: RecyclerView적용
