+++ app/src/main/java/woowacourse/movie/ui/home/MovieHomeActivity.kt
@@ -1,15 +1,16 @@
 package woowacourse.movie.ui.home
 
 import android.os.Bundle
-import android.widget.ListView
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.model.data.MovieContentsImpl
 import woowacourse.movie.model.movie.MovieContent
 import woowacourse.movie.ui.base.BaseActivity
-import woowacourse.movie.ui.home.adapter.MovieContentListAdapter
+import woowacourse.movie.ui.home.adapter.MovieContentAdapter
 
 class MovieHomeActivity : BaseActivity<MovieHomeContract.Presenter>(), MovieHomeContract.View {
-    private val movieContentList: ListView by lazy { findViewById(R.id.movie_content_list) }
+    private val movieContentList: RecyclerView by lazy { findViewById(R.id.movie_content_list) }
 
     override fun initializePresenter() = MovieHomePresenter(this, MovieContentsImpl)
 
@@ -21,6 +22,7 @@ class MovieHomeActivity : BaseActivity<MovieHomeContract.Presenter>(), MovieHome
     }
 
     override fun showMovieContents(movieContents: List<MovieContent>) {
-        movieContentList.adapter = MovieContentListAdapter(this, movieContents)
+        movieContentList.layoutManager = LinearLayoutManager(this)
+        movieContentList.adapter = MovieContentAdapter(movieContents)
     }
 }
