+++ app/src/main/res/layout/activity_movie_home.xml
@@ -7,11 +7,11 @@
     android:layout_height="match_parent"
     tools:context=".ui.home.MovieHomeActivity">
 
-    <ListView
+    <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/movie_content_list"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        tools:listitem="@layout/item_movie_content"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        tools:listitem="@layout/item_movie_content" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
